mnist/requirements.txt
Flask
tensorflow==2.8.0
numpy
Pillow
